add exceptions

+++ src/main/java/com/sims/product/ProductRepository.java
@@ -2,5 +2,5 @@ package com.sims.product;
 
 import com.sims.jpa.Jpa8Repository;
 
-public interface ProductRepository extends Jpa8Repository<Product, Long> {
+public interface ProductRepository extends Jpa8Repository<Product, Integer> {
 }
