+++ src/main/java/com/sims/warehouse/WarehouseRepository.java
@@ -2,5 +2,5 @@ package com.sims.warehouse;
 
 import com.sims.jpa.Jpa8Repository;
 
-public interface WarehouseRepository extends Jpa8Repository<Warehouse, Long> {
+public interface WarehouseRepository extends Jpa8Repository<Warehouse, Integer> {
 }
